refactor: move channels_v2/email_channel.py to channels/email_channel.py
