+++ apps/channels/channels_v2/registry.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from apps.channels.api_channel import ApiChannel
-from apps.channels.channels_v2.email_channel import EmailChannel
 from apps.channels.channels_v2.facebook_channel import FacebookMessengerChannel
 from apps.channels.channels_v2.slack_channel import SlackChannel
 from apps.channels.channels_v2.sureadhere_channel import SureAdhereChannel
@@ -11,6 +10,7 @@ from apps.channels.channels_v2.telegram_channel import TelegramChannel
 from apps.channels.channels_v2.web_channel import WebChannel
 from apps.channels.channels_v2.whatsapp_channel import WhatsappChannel
 from apps.channels.connect_channel import CommCareConnectChannel
+from apps.channels.email_channel import EmailChannel
 from apps.channels.models import ChannelPlatform
 
 if TYPE_CHECKING:
